Refactor and split code into classes
