+++ src/Sidebar.blp
@@ -1,0 +1,131 @@
+using Gtk 4.0;
+using Adw 1;
+
+template $Sidebar : Adw.NavigationPage {
+  child: Adw.ToolbarView {
+    [top]
+    Box {
+      styles ["toolbar"]
+
+      SearchEntry search_entry {
+        hexpand: true;
+        placeholder-text: _("Search Manuals");
+      }
+
+      MenuButton button_menu {
+        menu-model: menu_app;
+        icon-name: "open-menu-symbolic";
+        tooltip-text: _("Main Menu");
+        primary: true;
+      }
+    }
+
+    content: Stack stack {
+      transition-type: none;
+
+      Adw.StatusPage status_page {
+        title: _("No Results Found");
+        description: _("Try a different search term");
+        icon-name: "loupe-symbolic";
+        styles ["compact"]
+      }
+    };
+  };
+}
+
+menu menu_app {
+  section {
+    item {
+      custom: "themeswitcher";
+    }
+  }
+
+  section {
+    submenu {
+      label: _("Bookmarks");
+
+      item {
+        label: _("Developer Documentation");
+        action: "app.open_uri";
+        target: "https://developer.gnome.org/documentation/index.html";
+      }
+
+
+      item {
+        label: _("Human Interface Guidelines");
+        action: "app.open_uri";
+        target: "https://developer.gnome.org/hig/";
+      }
+
+      item {
+        label: _("JavaScript");
+        action: "app.open_uri";
+        target: "https://gjs.guide";
+      }
+
+      item {
+        label: _("Vala");
+        action: "app.open_uri";
+        target: "https://valadoc.org/";
+      }
+
+      item {
+        label: _("Rust");
+        action: "app.open_uri";
+        target: "https://gtk-rs.org";
+      }
+
+      item {
+        label: _("Python");
+        action: "app.open_uri";
+        target: "https://pygobject.readthedocs.io/";
+      }
+
+      item {
+        label: _("Blueprint");
+        action: "app.open_uri";
+        target: "https://jwestman.pages.gitlab.gnome.org/blueprint-compiler/";
+      }
+
+      item {
+        label: _("Discourse");
+        action: "app.open_uri";
+        target: "https://discourse.gnome.org/c/platform/5";
+      }
+    }
+
+    submenu {
+      label: _("Platform Tools");
+
+      item {
+        label: _("Adwaita Demo");
+        action: "app.platform_tools";
+        target: "adwaita-1-demo";
+      }
+
+      item {
+        label: _("GTK Demo");
+        action: "app.platform_tools";
+        target: "gtk4-demo";
+      }
+
+      item {
+        label: _("GTK Widget Factory");
+        action: "app.platform_tools";
+        target: "gtk4-widget-factory";
+      }
+    }
+  }
+
+  section {
+    item {
+      label: _("Keyboard Shortcuts");
+      action: "app.shortcuts";
+    }
+
+    item {
+      label: _("About Manuals");
+      action: "app.about";
+    }
+  }
+}
