+++ src/window.blp
@@ -3,101 +3,16 @@ using Adw 1;
 using GObject 2.0;
 using WebKit 6.0;
 
-Adw.Window documentation_viewer {
+template $ManualsWindow : Adw.Window {
   width-request: 400;
   height-request: 400;
   default-width: 1024;
   default-height: 768;
-  hide-on-close: true;
   title: bind webview.title;
 
   content: Adw.NavigationSplitView split_view {
-    sidebar: Adw.NavigationPage sidebar {
-      title: "";
-      child: Adw.ToolbarView {
-        [top]
-        Box {
-          styles ["toolbar"]
-
-          SearchEntry search_entry {
-            hexpand: true;
-            placeholder-text: _("Search Manuals");
-          }
-
-          MenuButton button_menu {
-            menu-model: menu_app;
-            icon-name: "open-menu-symbolic";
-            tooltip-text: _("Main Menu");
-            primary: true;
-          }
-        }
-
-        content: Stack stack {
-          transition-type: none;
-
-          ScrolledWindow browse_page {
-            vexpand: true;
-            width-request: 400;
-
-            ListView browse_list_view {
-              enable-rubberband: false;
-              factory: BuilderListItemFactory {
-                template ListItem {
-                  child: TreeExpander expander {
-                    list-row: bind template.item;
-                    child: Inscription {
-                      hexpand: true;
-                      nat-chars: 10;
-                      text-overflow: ellipsize_end;
-                      text: bind expander.item as <$DocumentationPage>.name;
-                    };
-                  };
-                }
-              };
-              styles ["navigation-sidebar"]
-            }
-          }
-
-          ScrolledWindow search_page {
-            vexpand: true;
-            width-request: 400;
-
-            ListView search_list_view {
-              enable-rubberband: false;
-              factory: BuilderListItemFactory {
-                template ListItem {
-                  child: Box {
-                    Inscription {
-                      valign: center;
-                      hexpand: true;
-                      nat-chars: 25;
-                      text-overflow: ellipsize_end;
-                      text: bind template.item as <$DocumentationPage>.search_name;
-                    }
-                    Button {
-                      valign: center;
-                      label: bind template.item as <$DocumentationPage>.tag;
-                      styles ["pill", "small", "doc-tag"]
-                    }
-                  };
-                }
-              };
-              styles ["navigation-sidebar"]
-            }
-          }
-
-          Adw.StatusPage status_page {
-            title: _("No Results Found");
-            description: _("Try a different search term");
-            icon-name: "loupe-symbolic";
-            styles ["compact"]
-          }
-        };
-      };
-    };
-
     content: Adw.NavigationPage {
-      title: bind documentation_viewer.title;
+      title: bind template.title;
       child: Adw.ToolbarView {
         [top]
         Adw.HeaderBar {
@@ -115,113 +30,8 @@ Adw.Window documentation_viewer {
             styles ["flat"]
           }
         }
-
-        WebKit.WebView webview {
-          vexpand: true;
-          settings: WebKit.Settings {
-            enable-back-forward-navigation-gestures: true;
-            enable-developer-extras: true;
-            enable-smooth-scrolling: true;
-          };
-        }
+        $WebView webview{}
       };
     };
   };
-}
-
-menu menu_app {
-  section {
-    item {
-      custom: "themeswitcher";
-    }
-  }
-
-  section {
-    submenu {
-      label: _("Bookmarks");
-
-      item {
-        label: _("Developer Documentation");
-        action: "app.open_uri";
-        target: "https://developer.gnome.org/documentation/index.html";
-      }
-
-
-      item {
-        label: _("Human Interface Guidelines");
-        action: "app.open_uri";
-        target: "https://developer.gnome.org/hig/";
-      }
-
-      item {
-        label: _("JavaScript");
-        action: "app.open_uri";
-        target: "https://gjs.guide";
-      }
-
-      item {
-        label: _("Vala");
-        action: "app.open_uri";
-        target: "https://valadoc.org/";
-      }
-
-      item {
-        label: _("Rust");
-        action: "app.open_uri";
-        target: "https://gtk-rs.org";
-      }
-
-      item {
-        label: _("Python");
-        action: "app.open_uri";
-        target: "https://pygobject.readthedocs.io/";
-      }
-
-      item {
-        label: _("Blueprint");
-        action: "app.open_uri";
-        target: "https://jwestman.pages.gitlab.gnome.org/blueprint-compiler/";
-      }
-
-      item {
-        label: _("Discourse");
-        action: "app.open_uri";
-        target: "https://discourse.gnome.org/c/platform/5";
-      }
-    }
-
-    submenu {
-      label: _("Platform Tools");
-
-      item {
-        label: _("Adwaita Demo");
-        action: "app.platform_tools";
-        target: "adwaita-1-demo";
-      }
-
-      item {
-        label: _("GTK Demo");
-        action: "app.platform_tools";
-        target: "gtk4-demo";
-      }
-
-      item {
-        label: _("GTK Widget Factory");
-        action: "app.platform_tools";
-        target: "gtk4-widget-factory";
-      }
-    }
-  }
-
-  section {
-    item {
-      label: _("Keyboard Shortcuts");
-      action: "app.shortcuts";
-    }
-
-    item {
-      label: _("About Manuals");
-      action: "app.about";
-    }
-  }
 }
